+++ src/npc/src/dialogue.c
@@ -112,66 +112,6 @@ int add_edge(convo_t *c, char *quip, char *from_id, char *to_id,
 }
 
 
-/**********************************************
- *       DIALOGUE EXECUTION FUNCTIONS         *
- **********************************************/
-
-/* Executes all actions on a node.
- *
- * Parameters:
- *  - n: node
- *  - game: the Chiventure game being run
- *
- * Returns:
- *  - SUCCESS if the action executed successfully, FAILURE if an error occured
- */
-int do_node_actions(node_t *n, game_t *game)
-{
-    node_action_t *cur_action = n->actions;
-
-    while (cur_action != NULL) {
-
-        switch(cur_action->action) {
-
-            case GIVE_ITEM: ;
-                npc_t *npc = get_npc_in_room(game->curr_room,
-                                             game->mode->mode_ctx);
-                item_t *item = get_item_in_hash(npc->inventory,
-                                                cur_action->action_id);
-                if (item == NULL) return FAILURE;
-                if (remove_item_from_npc(npc, item) != SUCCESS) return FAILURE;
-                if (add_item_to_player(game->curr_player, item) != SUCCESS)
-                    return FAILURE;
-                break;
-
-            case TAKE_ITEM:
-                npc = get_npc_in_room(game->curr_room, game->mode->mode_ctx);
-                item = get_item_in_hash(game->curr_player->inventory,
-                                                cur_action->action_id);
-                if (item == NULL) return FAILURE;
-                if (remove_item_from_player(game->curr_player, item) != SUCCESS)
-                    return FAILURE;
-                if (add_item_to_npc(npc, item) != SUCCESS) return FAILURE;
-                break;
-
-            case START_QUEST:
-                // to do
-                break;
-
-            case START_BATTLE:
-                // to do
-                break;
-
-            default:
-                return FAILURE;
-        }
-
-        cur_action = cur_action->next;
-    }
-
-    return SUCCESS;
-}
-
 /* Checks each edge and updates their availability status, then counts the
  * total number of available edges.
  *
@@ -278,97 +218,6 @@ char *create_return_string(node_t *n, int is_leaf)
 
     return buf;
 }
-
-/* See dialogue.h */
-char *start_conversation(convo_t *c, int *rc, game_t *game)
-{   
-    if (c == NULL) {
-        *rc = -1;
-        return NULL;
-    }
-
-    char *ret_str;
-
-    // Step 1: Set current node to first node
-    c->cur_node = c->all_nodes->node;
-
-    // Step 2: Execute actions (item, quest, battle, etc.), if any
-    if (do_node_actions(c->cur_node, game) != SUCCESS) {
-        *rc = -1;
-        return NULL;
-    }
-
-    // Step 3: Recheck the availability of each edge, count total avail. edges
-    if (update_edge_availabilities(c->cur_node) != SUCCESS) {
-        *rc = -1;
-        return NULL;
-    }
-
-    // Step 4: Prepare return code and return string
-    if (c->cur_node->num_available_edges == 0) *rc = 1;
-    else *rc = 0;
-
-    ret_str = create_return_string(c->cur_node, *rc);
-    if (ret_str == NULL) *rc = -1;
-    
-    return ret_str;
-}
-
-/* See dialogue.h */
-char *run_conversation_step(convo_t *c, int input, int *rc, game_t *game)
-{   
-    if (input > c->cur_node->num_available_edges)
-        input = c->cur_node->num_available_edges;
-
-    char *ret_str;
-    edge_list_t *cur_edge;
-
-    // Step 1: Traverse to the player's selected node
-    cur_edge = c->cur_node->edges;
-    // The logic here works as follows:
-    // (1) We only decrement input when we encounter an available edge, until
-    //     input become 1
-    // (2) However, we could end up on an unavailable edge, which is where
-    //     "|| cur_edge->availability != EDGE_AVAILABLE" comes in
-    // (3) Overall, this code ensures that we arrive at the player's selected
-    //     edge
-    while (input > 1 || cur_edge->availability != EDGE_AVAILABLE) {
-        if (cur_edge->availability == EDGE_AVAILABLE) input--;
-        cur_edge = cur_edge->next;
-    }
-
-    c->cur_node = cur_edge->edge->to;
-
-    // Step 2: Disable edge if the node it leads to has an action
-    // NOTE: This is a temporary solution that prevents issues like being able
-    //       to receive multiple copies of items, starting the same quest twice.
-    //       This SHOULD be changed / made more complex in the future.
-    if (c->cur_node->actions != NULL) {
-        cur_edge->availability = EDGE_DISABLED;
-    }
-
-    // Step 3: Execute actions (item, quest, battle, etc.), if any
-    if (do_node_actions(c->cur_node, game) != SUCCESS) {
-        *rc = -1;
-        return NULL;
-    }
-
-    // Step 4: Recheck the availability of each edge, count total avail. edges
-    if (update_edge_availabilities(c->cur_node) != SUCCESS) {
-        *rc = -1;
-        return NULL;
-    }
-
-    // Step 5: Prepare return code and return string
-    if (c->cur_node->num_available_edges == 0) *rc = 1;
-    else *rc = 0;
-
-    ret_str = create_return_string(c->cur_node, *rc);
-    if (ret_str == NULL) *rc = -1;
-
-    return ret_str;
-}
-
 
 /**********************************************
  *             ACTION FUNCTIONS               *
